chore: add pipenv dependency and lock
Adding pipenv as a dependency to the pipenv environment itself is admittedly
counter-intuitive, however this gives the advantage of being able to install all
dependencies (including dev-dependencies) via the common and generally available
`pip` package manager interface. The generation of requirements.txt and
requirements-dev.txt is abstracted through the Makefile.